Add files via upload
How do Londoners get around? Transport for London (TfL) is a vast public transport network that allows London's residents to efficiently travel around the UK's capital, to the tune of over 1.5 million journeys per day!

I worked with a Snowflake database containing data on journeys and transport types in London between 2010 and 2022. I wrote SQL queries to find the most popular transport methods, examine when the Emirates Airline cable car (which connects London's Royal Docks, where the Mayor's office is located, to Greenwich Peninsula, home of the O2 arena) was particularly busy, and identify rare periods when the Underground (also known as "the tube" to locals) was less busy.